Correção de Cardview -Correção na apresentação de dinheiro. -dinheiro descontado na carteira. -correção na apresentação da data de criação do serviço.

+++ Usuario/app/src/main/java/com/outsource/inovaufrpe/usuario/carteira/gui/AdicionarMoedaActivity.java
@@ -57,7 +57,7 @@ public class AdicionarMoedaActivity extends AppCompatActivity {
             public void onDataChange(DataSnapshot dataSnapshot) {
                 Usuario usuario = dataSnapshot.getValue(Usuario.class);
                 God moeda = new God(usuario.getCarteira());
-                moeda.adicionar(Long.valueOf(etAdicionarMoedaID.getText().toString()));
+                moeda.adicionar(Double.valueOf(etAdicionarMoedaID.getText().toString()));
                 usuario.setCarteira(moeda.getMoeda());
                 databaseReference.child("usuario").child(firebaseAuth.getCurrentUser().getUid()).setValue(usuario);
             }
