+++ Usuario/app/src/main/java/com/outsource/inovaufrpe/usuario/solicitante/gui/CadastroActivity.java
@@ -70,7 +70,7 @@ public class CadastroActivity extends Activity {
                             usuario.setSobrenome(etSobrenome.getText().toString());
                             usuario.setEmail(etEmail.getText().toString());
                             usuario.setTelefone(etTelefone.getText().toString().trim());
-                            usuario.setCarteira(Long.valueOf("0"));
+                            usuario.setCarteira(Double.valueOf("0"));
                             firebase.getUsuarioReference().child(user.getUid()).setValue(usuario);
                             startActivity(new Intent(CadastroActivity.this, MainActivity.class));
                             finish();
